backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
langchain-core==0.1.52
langchain-community==0.0.38
langchain-google-genai==1.0.1
google-generativeai==0.3.2
pymysql==1.1.0
